Utils changes from the React template has moved to the vis-core package
