+++ packages/vis-core/src/utils/config.test.js
