+++ packages/vis-core/src/utils/imageLoader.js
