--- packages/vis-core/src/utilsNew/copyright.js
@@ -1,6 +1,0 @@
-export const buildCopyright = (extraHtml) => {
-  const year = new Date().getFullYear();
-  return `Contains OS data © Crown copyright and database right ${year}${
-    extraHtml ? ` | ${extraHtml}` : ''
-  }`;
-};
