--- packages/vis-core/src/utilsNew/index.js
@@ -1,12 +1,0 @@
-export * from './apiParams'
-export * from './config'
-export * from './map'
-export * from './math'
-export * from './state'
-export * from './text'
-export * from './colorSchemes'
-export * from './syncMaps'
-export * from './validation'
-export * from './nav'
-export * from './dataFormatter'
-export * from './dynamicStyling'
